wifi get all wifi endpoint success
