+++ poc/server/app/__init__.py
@@ -3,6 +3,7 @@ from flask import Blueprint
 
 from .main.controller.user_controller import api as user_ns
 from .main.controller.auth_controller import api as auth_ns
+from .main.controller.wifi_data_controller import api as wifi_ns
 
 blueprint = Blueprint('api', __name__)
 authorizations = {
@@ -24,3 +25,4 @@ api = Api(
 
 api.add_namespace(user_ns, path='/user')
 api.add_namespace(auth_ns)
+api.add_namespace(wifi_ns, path='/wifi')
