+++ poc/server/app/main/model/wifi_data.py
