+++ poc/server/requirements.txt
@@ -4,7 +4,7 @@ Flask==2.2.3
 itsdangerous==2.2.0
 Jinja2==3.1.3
 MarkupSafe==2.1.5
-Werkzeug==3.0.2
+Werkzeug==2.2.2
 alembic==1.9.4
 aniso8601==9.0.1
 attrs==23.2.0
